towards leaner react lib definitions
Summary: It turns out that it's not very useful to keep State around in the type definitions of ReactElement and ReactClass (even though it is useful in ReactComponent): intuitively, once a definition is checked, nobody cares what State a React class carries. Now State is "deprecated" (although not removed, since it would provoke a huge number of arity-mismatch errors where type args were explicitly passed in): cleanup should follow in a separate diff.

Reviewed By: samwgoldman

Differential Revision: D2924975

fb-gh-sync-id: 215c889de99f98b044cd894eef95afd11fe3524d
shipit-source-id: 215c889de99f98b044cd894eef95afd11fe3524d