Fix `import {SomeTypeAlias} from ...` to give proper error
Summary:Previously if you did the following:

```
// types.js
export type foo = number;
```

```
// main.js
import {foo} from "./types";
```

You would get no errors (notice the non-type import of a type alias). This is because we didn't validate that type-imports are actually types.

This diff fixes this (and, in the process, also cleans up a bunch of import inference logic)

Reviewed By: samwgoldman

Differential Revision: D2906636

fb-gh-sync-id: e5776fa2ac9ebe7422e805bb052268fa32921c14
shipit-source-id: e5776fa2ac9ebe7422e805bb052268fa32921c14