Cover more cases for hopping through require()s with get-def
Summary:https://github.com/facebook/flow/commit/3edcf6e5c815faf385058638cd92fe0600c77b03 made it possible, given `var {Foo} = require('Foo');` to get-def on `{Foo}` so that it hops through `require('Foo')` and all the way to the actual property on the definition side via get-def.

It didn't, however: (a) cover the simpler case of `var Foo = require('Foo').Foo` and (b) improve the half-assed location we give when calling get-def on a `require()` (we were giving a "none" location for the module which only had the module's file path (but crappy line/col info)).

This diff fixes (a) and gives a useful line/col for (b) by tracing back to the point where module.exports is assigned. Fixing (b) is especially useful if you're trying to trace a definition through your program (via repeated get-def jumps in an editor, for example).

(Note that this cleanup work before adding a `--jump-through-imports` flag per #9853750)

Reviewed By: samwgoldman

Differential Revision: D2914916

fb-gh-sync-id: 8c1b7c391fc47346fbc5b9d7a16af10c8be0adce
shipit-source-id: 8c1b7c391fc47346fbc5b9d7a16af10c8be0adce