[hack] Fix OSS build after dynamic yield error message fix
Summary:The dynamic yield behavior message was modified, but one of the strings
for the removed `%s` forgot to be removed as well.

We ignore this warning internally, but it appears in OSS.

```
File "typing/typing_dynamic_yield.ml", line 57, characters 48-216:
Error: This expression has type
         (string -> string -> 'a -> 'b, unit, string, string, string,
          'a -> 'b)
         format6
       but an expression was expected of type
         (string -> string -> 'a -> 'b, unit, string) format =
           (string -> string -> 'a -> 'b, unit, string, string, string,
            string)
           format6
       Type 'a -> 'b is not compatible with type string
Command exited with code 2.
```

Differential Revision: D2942102

fb-gh-sync-id: 1dceec3f301485d392c3ff33158bda6fd0fd50d9
shipit-source-id: 1dceec3f301485d392c3ff33158bda6fd0fd50d9