[hack] Fix bugs in Marshal_tools limiting message size from 2^31 to 2^16
Summary:I need to pass files_info structure from typechecker process to IDE process.

I initially wanted to use global storage for that, but int3 told me that if we are not solving "single writer, multiple readers" problem there is no advantage and we can just send it through channel.

files_info is bigger than all of the message types we send so far[1], which revealed two bugs in Marshal_tools:

 - the code used single Unix.read call which limits the size of passed message to system buffer size (2^16)
 - of-by-one mistake limited the number of usable bytes in preamble to 3, mangling messages with size bigger than 2^24

Reviewed By: int3, alexchow

Differential Revision: D2947605

fb-gh-sync-id: 7389f80e2edbfb6ef71c29685bd97f7fa578ec66
shipit-source-id: 7389f80e2edbfb6ef71c29685bd97f7fa578ec66