Add parsing support for default values in destructuring declarations
Summary:This diff creates a new code path for destructuring declarations. This does include variable declarations (`let {x:y=z} = e`) and function param lists (`function f([a=0]) {}`), but does not include assignment expressions (`({x:y=z} = e)`), which are not only a mess of backtracking, but also permit destructuring reference bindings[0], which need special care.

Furthermore, this diff only adds parser support and does not change the current destructuring behavior within Flow. That is, we will still display "unsupported destructuring" when we encounter a default.

I left the "convert expression to pattern" code, which is still used for the assignment expression case. I hope to replace this with a proper solution that permits both defaults and reference bindings in subsequent diffs.

Misc:
* Updated test which was a parse error we wrongly interpreted as destructuring to what I suspect was intended.

0. https://gist.github.com/DmitrySoshnikov/79ee5e9779062757de4b

Reviewed By: gabelevi

Differential Revision: D2943578

fb-gh-sync-id: ab801ff923c62030377c97963247657cd2b669bf
shipit-source-id: ab801ff923c62030377c97963247657cd2b669bf