Add typing support for default values in destructuring declarations
Summary:A default in a destructuring pattern causes a lower bound to flow into the tvar
of the introduced binding. Because destructuring is recursive and defaults can
be provided at every level, I co-opted some of the DestructuringT machinery
(specifically, `eval_selector`) to convert multi-layered defaults into a union.
See the added `Default` module for more specifics.

This adds yet another parameter to the `destructuring` helper: `default`. I
also renamed `expr` to `init` for clarity.

Reviewed By: avikchaudhuri

Differential Revision: D2948894

fb-gh-sync-id: f94728ec1bc15c0e09d68b8673f0de501fa012c9
shipit-source-id: f94728ec1bc15c0e09d68b8673f0de501fa012c9