[hack] Remove TypedefPosHeap
Summary:Previously, when a typedef got declared, it was added to both the
ClassPosHeap and the TypedefHeap. Adding it to the ClassPosHeap was
necessary because classes and typedefs both live in the same namespace,
so declaring a class and a typedef of the same name was an error that we
wanted the typechecker to be able to identify. However, instead of
adding a duplicate entry into the TypedefPosHeap to remember that a
given name was a typedef, we can instead store the kind of the
name in the same ClassPosHeap (ClassIdHeap) table.

This makes lazy decl a bit cleaner -- when a hint is missing a
corresponding declaration, instead of having to query both ClassPosHeap
and TypedefPosHeap to figure out if and how it needs to be declared, we
can just substitute one query to ClassIdHeap.

I also unified the no_typedef check with the rest of the hint-checking logic in
naming. Since we no longer require a separate hashtable query it didn't make
sense to separate it. This revealed a number of places where we should have
been checking for non-typedefs but weren't. Tests were added to demonstrate
this.

Reviewed By: dlreeves

Differential Revision: D2915262

fb-gh-sync-id: 3aeb65e13d8677769070aba8836736dd4c43585e
shipit-source-id: 3aeb65e13d8677769070aba8836736dd4c43585e