Add React docs
Summary:Added a suppression comment and changed the docs generation to exclude the
suppression comments. The errors are still generated for the docs, but a
"passing" docs set should have no errors.

One of the examples uses esproposal static and instance fields, which I turned
on in the .flowconfig, but since we use check-contents and check-contents
doesn't support .flowconfig nor has command line options for those flags, the
generated docs have errors. :(

I noticed that lib/react.js had a nonsense declaration for getDefaultProps, but
it doesn't seem to cause any errors in practice. Still worth changing.

Reviewed By: avikchaudhuri

Differential Revision: D2958285

fb-gh-sync-id: 58bb790d3762dfe29dc871864373f580c1581110
shipit-source-id: 58bb790d3762dfe29dc871864373f580c1581110