use travis to upload release binaries
Summary:when a commit is tagged, travis will automatically build the .zip, create the release on GitHub if it doesn't exist, upload the binary and publish the release.

aside from it being cool to automate this, it needs to happen before the site gets rebuilt so that the site pulls the correct release.

Reviewed By: samwgoldman

Differential Revision: D2958553

fb-gh-sync-id: cc3061e1dddc71f89e2cac2712e875900d335f38
shipit-source-id: cc3061e1dddc71f89e2cac2712e875900d335f38