refine falsy values to literals
Summary:the only falsy string is `""` and the only falsy number is `0`, so we can get rid of the `Falsy` variant and just use `Literal`.

This has the bonus of fixing any rules that were matching literal values against inferred-falsy strings, like in the issue below.

Fixes https://github.com/facebook/flow/issues/1427

Reviewed By: samwgoldman

Differential Revision: D2958656

fb-gh-sync-id: 5c7c690256357acf979310286e6ba6813d012ff2
shipit-source-id: 5c7c690256357acf979310286e6ba6813d012ff2