[docs] convert index.md to JS, update error output
Summary:Adds inline errors and updates the output to use our "new" error format.

Fixes a bug in the generator so that we don't strip all leading whitespace from each line of a comment; rather, only as much as the least-indented line is indented.

Also tweaked some CSS to make the Flow- and rouge-highlighted code blocks indented the same.

Reviewed By: bhosmer

Differential Revision: D2959602

fb-gh-sync-id: 333275e49314b9a764dd187f231b3bddee0ec399
shipit-source-id: 333275e49314b9a764dd187f231b3bddee0ec399