Add Refinement docs
Summary:I mostly copied this from out existing dynamic type test docs. I'm sure there's
more to say on this topic, but for now it's a win to have this be type checked.

In fact, I realized that one of the examples under caveats is no longer a
caveat, thanks to read/write sets.

Reviewed By: bhosmer

Differential Revision: D2958700

fb-gh-sync-id: 755cd7ee8fb2ba816836d2f167eaf80ef7cd4891
shipit-source-id: 755cd7ee8fb2ba816836d2f167eaf80ef7cd4891