[flow docs][handbook] Modules
Summary:Docs for modules.

Still left TODO: CommonJS <-> ES Module interop section

NOTE: There are a lot of recorded errors in the examples here because (a) referencing modules that don't exist and (b) [necessarily] duplicate bindings being created. My idea to tackle this is to hack in some delimiter syntax that lets you specify that an example section should be broken out and treated as if it's a separate file during the build. I'll come back to this at some point later (if someone doesn't beat me to it).

Reviewed By: samwgoldman

Differential Revision: D2959477

fb-gh-sync-id: 5f4b31678b6f1739af17a887288db4d2027a5605
shipit-source-id: 5f4b31678b6f1739af17a887288db4d2027a5605