[parser] "fix" location of JSXEmptyExpression
Summary:Fixes a discrepancy between us and both babel and esprima, where the location of `JSXEmptyLocation` includes the `{` and `}`.

Given `(<div>{}</div>)`, both babel and esprima return this:

```
          {
            "type": "JSXExpressionContainer",
            "start": 6,
            "end": 8,
            "range": [
              6,
              8
            ],
            "expression": {
              "type": "JSXEmptyExpression",
              "start": 7,
              "end": 7,
              "range": [
                7,
                7
              ]
            }
          }
```

but we were returning [6, 8] for the location of the empty expression.

Thanks to ariya for noting this in https://github.com/jquery/esprima/pull/1470

Reviewed By: gabelevi

Differential Revision: D2946420

fb-gh-sync-id: 8d8f8a34612fd30b89fe69fe453077e5b990d9d3
shipit-source-id: 8d8f8a34612fd30b89fe69fe453077e5b990d9d3