Support JSON files with toplevel arrays
Summary:There are plenty of JSON files in the wild (on npm, etc) that start with an array at the toplevel. This is a source of some moderate number of JSON file parsing errors that I've seen.

This fixes the json-parser function to handle both toplevel objects and arrays.
Closes https://github.com/facebook/flow/pull/1436

Reviewed By: mroch

Differential Revision: D2966182

Pulled By: jeffmo

fb-gh-sync-id: b952713a3c74cc77ebb7eb746825e10b970cd66a
shipit-source-id: b952713a3c74cc77ebb7eb746825e10b970cd66a