[flow docs] Add `$DocIssue`, bring docs errors to 0, and fix check-contents to respect suppression_comment
Summary:`$DocIssue` is a new suppression comment that (a) suppresses errors and (b) hides the errors from the UI (because the errors are legit, but aren't meant to be a part of the documentation).

Note that `check-contents` wasn't filtering suppressed errors in it's output, so this diff also fixes this.

Reviewed By: mroch

Differential Revision: D2967125

fb-gh-sync-id: ad29308092daf9c86761d3f6f74a91b3f762a16e
shipit-source-id: ad29308092daf9c86761d3f6f74a91b3f762a16e