import 'react' fix
Summary:There was a race condition in the way defaultProps and config are merged before props checking, causing a React class with no defaultProps to skip config and present an empty object for props checking. What sucks is that this race condition is a known problem with ObjAssignT, and was fixed, but the fix was not used for this particular operation.

This bug triggers when an explicit import 'react' used to construct a React class with no defaultProps and an implicit require 'react' used to construct a jsx appear in the same file, causing two different constraints to hang from the same variable and get processed in an unfortunate sequence.

Reviewed By: jeffmo

Differential Revision: D2974368

fb-gh-sync-id: e05ad91879a33f9d04693bc64821140342db5f6e
shipit-source-id: e05ad91879a33f9d04693bc64821140342db5f6e