[hack] Shuffle around and improve quick reference docs
Summary:The quick reference was probably our best piece of documentation, but it was
pretty long. Also, some parts of it were less useful than others, like the
section on Boolean, String, and Number wrapper objects.

The previous quick reference was serving two purposes, really:
1. Explaining the built-in types
2. Giving a high-level overview of core functionality

To add a bit of focus, I added a new section, Built-in Types which follows the
quick reference and goes into more depth on that topic. Also, I moves some
examples to the Array, Object, and Class documentation.

The goal is that the quick reference should be easy to read top-to-bottom to get
a sense of what we're all about.

Reviewed By: bhosmer

Differential Revision: D2959091

fb-gh-sync-id: 215696763c72afb641e788d642b1d0de4ab9f411
shipit-source-id: 215696763c72afb641e788d642b1d0de4ab9f411