[hack] Remove weak docs
Summary:These three sections, Variables, Operators, and Primitives, were not very
informative, and I want to remove them in order to make the more effective docs
more prominent.

The content in Variables will be supplanted by forthcoming docs describing
"Syntax" which will cover the same content and much more.

The information in Operators was sorta kinda useful, but I don't think that
people who need that information would know how to find it. A better place for
this kind of documentation is more of a language spec, which shows the typing
rules for expressions. Until we have that, I think Flow is clear enough about
what it infers.

The Primitives docs seems clearly supplanted by the new Built-in Types doc.

I didn't break the URLs. They still go to the old docs, but I removed them from
nav and updated the links between pages.

Reviewed By: jeffmo

Differential Revision: D2983867

fb-gh-sync-id: 4ab660e2d7ce7b1af0f4fdb0934fb48daf2ad9a2
shipit-source-id: 4ab660e2d7ce7b1af0f4fdb0934fb48daf2ad9a2