[hack] module.name_mapper.extension
Summary:A huge use-case for `module.name_mapper` is to just stub out CSS and image files for things like CSS Modules and image loader. For this, having to write a full-on regexp is a little annoying (and easy to mess up).

This just adds a simpler version of `module.name_mapper` that takes a file extension as the input rather than a full regular expression.

Reviewed By: samwgoldman

Differential Revision: D2961357

fb-gh-sync-id: 700d779006a19ada9c7f2ebae81d2d5d8ceeface
shipit-source-id: 700d779006a19ada9c7f2ebae81d2d5d8ceeface