[hack] make sure directory exists before creating lock/log files
Summary:given a filename, `Lock.grab`, `PidFile.init`, and the function that opens the log file for writing all expect the directory to exist so that they can create the file.

at the same time, we currently create the directory when we generate the filename, even if we're just going to see if the file exists (e.g. `Lock.check`). clearly the file doesn't exist if its directory doesn't exist, so removing the mkdir from flowConfig.ml  makes fewer unnecessary directories.

Reviewed By: int3

Differential Revision: D2981713

fb-gh-sync-id: 0aa810993ab8bfcbfcc1c19ca1968925e609cd84
shipit-source-id: 0aa810993ab8bfcbfcc1c19ca1968925e609cd84