[hack] hacky bandaid for race condition in integration test
Summary:this test is racy. if the server hasn't been notified of the `mv` by the time `flow status` runs, then the test fails.

adding a `sleep` gives it a chance to hopefully be notified. this is still racy in theory, and 1 second is probably way longer than it needs to be, but I want to keep the `sleep` command portable and fractional seconds are non-standard.

I think a better way to do this is to have a command or flag that blocks until an update is received. it could be `flow status --wait`, or a separate command like `dfind --wait`, which we could run in a subshell before the `mv` and then `wait` for it to exit after the `mv`, then call `flow status` like normal (this isolates the waiting stuff to the tests). All of that is future work, though... I just want to stabilize the tests ASAP.

Reviewed By: avikchaudhuri

Differential Revision: D2992850

fb-gh-sync-id: 2bd9143dd2e7903c38b3b987071ae70d0ccf5041
shipit-source-id: 2bd9143dd2e7903c38b3b987071ae70d0ccf5041