Change AST for type parameters, include variance
Summary:Before this change, we re-used the Identifier AST internally, and output just an Identifier node from the estree translation.

This commit replaces the internal AST with a record specifically for type parameters, it also changes the JSON AST output to a) include variance and b) put the bound under a key "bound" instead of "typeAnnotation" which was residue from the Identifier overloading.

Reviewed By: bhosmer

Differential Revision: D2993128

fb-gh-sync-id: 664e1af1a95b60c7e4576df57a84c8a5fe31102c
shipit-source-id: 664e1af1a95b60c7e4576df57a84c8a5fe31102c