[flow docs] Main page, "Getting Started", & "Five Simple Examples" revamp
Summary:Per title, revamps these three pages to flow neatly between each other.

Focus is on "new" user visiting the site (but also added a small link to the "Quick Reference" on the main page for those coming back for reference)

Additionally, I added a new magic doc comment: `$DocHide` which will hide the line it is on from being rendered in the UI (but not hide it from flow when running `flow check website`).

Reviewed By: mroch

Differential Revision: D3008053

fb-gh-sync-id: f06b88cd2869e80bddfe6e0f74b6b0c0ae7e708c
shipit-source-id: f06b88cd2869e80bddfe6e0f74b6b0c0ae7e708c