Fix broken link to React example.
Summary:This pull request updates the broken link to the React example in the documentation. It might be worthwhile to run [html-proofer](https://github.com/gjtorikian/html-proofer) on Travis builds to avoid broken links in the documentation in the future. I am more than happy to take a look at it if it sounds useful.

Fixes #1494.
Closes https://github.com/facebook/flow/pull/1496

Differential Revision: D3017232

Pulled By: jeffmo

fb-gh-sync-id: 921f0e65ee69f0d9834d3e7f82e993e37f01debf
shipit-source-id: 921f0e65ee69f0d9834d3e7f82e993e37f01debf