change structure of location JSON
Summary:changes the JSON format for locations from

```
// <= 0.22.1 (current release)
{
  // other props, like the error message
  "path": "foo/bar.js",
  "line": 1,
  "endline": 1,
  "start": 12,
  "end": 15
}

// master
{
  // other props, like the error message
  "loc": {
    "path": "foo/bar.js",
    "line": 1,
    "endline": 1,
    "start": 12,
    "end": 15
  }
}
```

to

```
{
  // other props
  "loc": {
    "source": "foo/bar.js",
    "start": {
      "line": 1,
      "column": 12,
      "offset": 12
    },
    "end": {
      "line": 1,
      "column": 15,
      "offset": 15
    }
  },
  // the rest is for backwards-compatibility
  "path": "foo/bar.js",
  "line": 1,
  ...
}
```

So, this does two things: it restores the props that were moved into the `loc` property in master, and changes the props under `loc` so that their line/col are structured differently. One reason for this is so that you can do range math like `distance(x, y) { return x.offset - y.offset }` without caring whether it's a start or end position (e.g. the length of a range (start to end), or the length of a gap between ranges (end to start)). This is also the format used in the AST.

It also leaves the old behavior intact since it doesn't confict, to give clients a bit of time to upgrade. We still reserve the right to remove this anytime.

Reviewed By: bhosmer

Differential Revision: D3022340

fb-gh-sync-id: fd22fb4bb0765a6fa4ac2d75228331b526a5dfd7
shipit-source-id: fd22fb4bb0765a6fa4ac2d75228331b526a5dfd7