Allow `Event` to be parameterized on its target
Summary:In places where people specify the type of an `Event` (usually the parameter of an event handler), it is common that the target of the event is fairly well known/implicitly expected (i.e.: `onInputChange(e: Event)` expects some kind of event on an `HTMLInputElement` -- developers usually infer this easily by context/naming/etc).

There is a common annoyance, however, when someone wishes to access the `target` of the event: They have to refine `e.target` explicitly in some kind of if-branch. This feels pretty heavyweight and is mostly just Flow boilerplate leaking into the handler's logic. So the idea here is to allow users to provide this hint to Flow in the Flow type itself rather than leaking into the logic in a way that feels heavy-handed and robotic.

So this logic:

```
...
  _onVideoChange(event: Event) {
    if (event.target instanceof HTMLVideoElement) {
      (event.target.videoWidth: number);
    }
  }
...
```

narrows to:

```
...
  _onVideoChange(event: Event<HTMLVideoElement>) {
    (event.target.videoWidth: number);
  }
...
```

Reviewed By: zpao

Differential Revision: D2927298

fb-gh-sync-id: 1576975089c3add4c5ce034ce4469b7b54fec484
shipit-source-id: 1576975089c3add4c5ce034ce4469b7b54fec484