Declare more built-ins with @@iterator, get rid of Array.from "any" sig
Summary:Overloads + any sigs are bad -- even if a different signature matches on the params, if the return value is mis-used then Flow lets it fall into the "any" signature. This basically means any overload with an "any" fallback has almost no typechecking other than that fallback.

This tightens up the signature to use mixed instead (which will force people to declare an explicit type if they are passing a sufficiently strange input to Array.from) and also tags more iterble things as iterable. This way they will be matched by the iterable signature.

Reviewed By: samwgoldman

Differential Revision: D3025463

fb-gh-sync-id: c9d8aa0f231b9185d8f0d8394b0ae561d7ce3a3f
shipit-source-id: c9d8aa0f231b9185d8f0d8394b0ae561d7ce3a3f