[hack] Functorize Naming
Summary:Naming is called both in the decl phase and type-check phase. In the
decl phase it's mostly used to construct things that do not belong in
function bodies; examples include classes, their member fields, and
global constants. This part of naming is entirely self-contained; it
only uses the data from the AST in the current file, and does not need
to cross-reference decl type data from other files.

In the type-check phase, Naming is invoked again, this time to name the
bodies of functions. Now it requires decl type data in order to know
which function calls are marked as `noreturn`, because this affects
which local variables are considered to be defined at the end of a
statement.

So decl depends on naming, but naming also depends on decl, creating
a circular dependency. The obvious solution would be to split it into
two, but this is nontrivial because decl-phase naming also does some
naming of expressions -- for example, constant initializers and default
parameter values have them. Of course, none of these expressions can
actually contain local variables, but our code is not written in a way
that the OCaml type system can understand that.

elgenie's solution (when he implemented `noreturn`) was to have naming
depend on the Typing_heap but not on the actual decl code. However, this
will no longer work with lazy decl, because we need the
type-check-phase-naming to be able to invoke the decl code if the
function isn't found in the heap.

As a hacky alternative, I'm parameterizing GetLocals so that it is a
no-op in the decl phase but can be properly instantiated with
Typing_get_locals in the type-check phase.

Also cleaned up a bazillion 80 char lint errors while I was at it (since `arc
lint` yelled at me for every line I touched...)

Reviewed By: dlreeves

Differential Revision: D3023579

fb-gh-sync-id: cff1ae879cb48a56d6c86266524b9e2515ffde56
shipit-source-id: cff1ae879cb48a56d6c86266524b9e2515ffde56